added llength

+++ src/main/scala/BIDMat/Benchmarks.scala
@@ -134,7 +134,7 @@ object AltaVista {
 	  	}
 	    Mat.ilexsort2(ii.data, jj.data)
 	    saveAs("/big/Yahoo/G2/test.mat",ii,"ii",jj,"jj")
-	    val avals = ones(ii.length, 1).data
+	    val avals = ones(ii.length.toInt, 1).data
 	    sofar = SparseMat.remdups(ii.data, jj.data, avals)
 	  	val ss = new SMat(maxv, maxv, sofar, SparseMat.incInds(jj.data), new Array[Int](maxv+1), avals)
 	    SparseMat.compressInds(ii.data, maxv, ss.jc, sofar)
